cleaning up, removed KeyType
