+++ packages/core/tests/capability/wnfs.test.ts
@@ -12,11 +12,6 @@ describe("wnfs public capability", () => {
 
   beforeAll(loadTestPlugins)
 
-  // @TODO undo
-  it('passes', () => {
-    expect(true)
-  })
-
   it("works with a simple example", async () => {
     const { ucan } = await makeSimpleDelegation(
       [ wnfsCapability("//boris.fission.name/public/Apps/", "OVERWRITE") ],
