+++ packages/core/tests/token.test.ts
@@ -2,7 +2,6 @@ import * as uint8arrays from "uint8arrays"
 
 import * as capability from "../src/capability"
 import * as token from "../src/token"
-// import { verifySignatureUtf8 } from "../src/did"
 import { loadTestPlugins } from "./setup.js"
 import { alice, bob } from "./fixtures"
 
@@ -169,16 +168,3 @@ describe("token.validate", () => {
     expect(token.isTooEarly(activeUcan)).toBe(false)
   })
 })
-
-// @TODO Move this??
-// describe("verifySignatureUtf8", () => {
-
-//   it("works with an example", async () => {
-//     const [ header, payload, signature ] = token.encode(await token.build({
-//       issuer: alice,
-//       audience: bob.did(),
-//     })).split(".")
-//     expect(await verifySignatureUtf8(`${header}.${payload}`, signature, alice.did())).toEqual(true)
-//   })
-
-// })
