+++ packages/plugins/tests/ecdsa.test.ts
@@ -1,17 +1,18 @@
-import * as did from "../src/did"
-import ECDSAKeyPair from "../src/keypair/ecdsa"
+import { p256Plugin } from "../src/p256/plugin.js"
+// import * as did from "../src/did"
+import EcdsaKeyPair from "../src/p256/keypair.js"
 
 
 describe("ecdsa", () => {
 
-  let p256Keypair: ECDSAKeyPair
+  let p256Keypair: EcdsaKeyPair
 
   let p256Signature: Uint8Array
 
   const data = new Uint8Array([1, 2, 3, 4, 5, 6, 7, 8, 9])
 
   it("creates an ecdsa keypairs with different curves", async () => {
-    p256Keypair = await ECDSAKeyPair.create()
+    p256Keypair = await EcdsaKeyPair.create()
   })
 
   it("returns a publicKeyStr and did: curve P-256", () => {
