+++ packages/ucans/src/index.ts
@@ -1,19 +1,7 @@
+import { loadPlugins } from "@ucans/core"
+import * as plugins from "@ucans/plugins"
+
+loadPlugins(plugins.defaults)
+
 export * from "@ucans/core"
-export * from "@ucans/defaults"
-
-
-
-// export * from "./attenuation.js"
-// export * from "./builder.js"
-// export * from "./did.js"
-// export * from "./keypair/ed25519.js"
-// export * from "./keypair/rsa.js"
-// export * from "./store.js"
-// export * from "./token.js"
-// export * from "./types.js"
-// export * from "./verify.js"
-
-// export * as keypair from "./keypair/index.js"
-// export * as capability from "./capability/index.js"
-
-// export { Capability, EncodedCapability, isCapability } from "./capability/index.js"
+export * from "@ucans/plugins"
